# Conflicts:
#	gradle/libs.versions.toml

+++ modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/schema/SqlSchemaManager.java
@@ -40,5 +40,5 @@ public interface SqlSchemaManager {
      *
      * @param version SQL schema version to wait.
      */
-    CompletableFuture<Void> schemaReadyFuture(long version);
+    CompletableFuture<Void> schemaReadyFuture(int version);
 }
